+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 * nostr: add `TagKind::custom` constructor ([Yuki Kishimoto])
 * nostr: add `Timestamp::from_secs` ([Yuki Kishimoto])
 * nostr: add `EventId::from_byte_array` ([Yuki Kishimoto])
+* nostr: add `Timestamp::min` and `Timestamp::max` ([Yuki Kishimoto])
 * database: add `DatabaseHelper::fast_query` ([Yuki Kishimoto])
 * database: add `NostrDatabase::check_event` ([Yuki Kishimoto])
 * relay-builder: add `MockRelay` ([Yuki Kishimoto])